Lay base for moving ShoppingList data into SQLite
I have made a SQLite db holding the basic data that goes into the
shopping list lines. At the moment i am not using the database as the dynamic data source (adding data to it/deleting data from it)
whenever i am modifying the list. Instead i extract data from the
database and put it in a ArrayList that is the dynamic data holder.